requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.2.1
djangorestframework==3.16.0
pillow==11.2.1
qrcode==8.2
sqlparse==0.5.3
tzdata==2025.2
